# backend/openapi.yml
openapi: 3.1.0
info:
  title: Fahrtenbuch API
  version: 1.0.0-dev
  description: |
    REST-API für das selbst gehostete Fahrtenbuch. Die API umfasst
    Authentifizierung, MFA, Passkeys, App-Pairing, Geräteverwaltung,
    Benutzer, Fahrzeuge, Fahrten, Tracking, Statistiken, Import/Export,
    Geocoding sowie zentrale Kartendienst-Konfiguration.
  license:
    name: GNU Affero General Public License v3.0
    identifier: AGPL-3.0-only
servers:
  - url: /api/v1
    description: Relativ zum Fahrtenbuch-Server
security:
  - bearerAuth: []
...
paths:
...
  /vehicles:
    get:
      tags: [Vehicles]
      summary: Fahrzeuge auflisten
      responses:
        '200':
          description: Fahrzeuge
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Vehicle' }
    post:
      tags: [Vehicles]
      summary: Fahrzeug anlegen
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VehicleInput' }
      responses:
        '201':
          description: Fahrzeug angelegt
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehicle' }
  /vehicles/{id}:
    get:
      tags: [Vehicles]
      summary: Fahrzeug abrufen
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '200':
          description: Fahrzeug
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehicle' }
    put:
      tags: [Vehicles]
      summary: Fahrzeug aktualisieren
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VehicleInput' }
      responses:
        '200':
          description: Fahrzeug aktualisiert
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vehicle' }
    delete:
      tags: [Vehicles]
      summary: Fahrzeug löschen
      parameters:
        - $ref: '#/components/parameters/Id'
      responses:
        '204': { description: Fahrzeug gelöscht }
  /vehicles/{id}/bluetooth:
    put:
      tags: [Vehicles]
      summary: Bluetooth-Zuordnung aktualisieren
      parameters:
        - $ref: '#/components/parameters/Id'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [bluetoothMac]
              properties:
                bluetoothMac:
                  type: [string, 'null']
                  pattern: '^([0-9A-Fa-f]{2}:){5}[0-9A-Fa-f]{2}$'
      responses:
        '200': { description: Bluetooth-Zuordnung aktualisiert }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  schemas:
...
    Vehicle:
      type: object
      required: [id, name, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        licensePlate: { type: [string, 'null'] }
        vin: { type: [string, 'null'] }
        odometerKm: { type: [number, 'null'], minimum: 0 }
        color: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        bluetoothMac: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    VehicleInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        licensePlate: { type: [string, 'null'] }
        vin: { type: [string, 'null'] }
        odometerKm: { type: [number, 'null'], minimum: 0 }
        color: { type: [string, 'null'] }
        notes: { type: [string, 'null'] }
        bluetoothMac:
          type: [string, 'null']
          pattern: '^([0-9A-Fa-f]{2}:){5}[0-9A-Fa-f]{2}$'
...